Refine test for GROOVY-9204
Use isolated test java classes instead.

(cherry picked from commit 29f43f165f002299e3e7e2f8e8daff003dbb3403)
