+++ src/test/groovy/bugs/groovy9204/Groovy9204.groovy
@@ -1,0 +1,49 @@
+/*
+ *  Licensed to the Apache Software Foundation (ASF) under one
+ *  or more contributor license agreements.  See the NOTICE file
+ *  distributed with this work for additional information
+ *  regarding copyright ownership.  The ASF licenses this file
+ *  to you under the Apache License, Version 2.0 (the
+ *  "License"); you may not use this file except in compliance
+ *  with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+package groovy.bugs.groovy9204
+
+import groovy.test.GroovyTestCase
+
+class Groovy9204 extends GroovyTestCase {
+    void testGenerics() {
+        assertScript '''
+            package groovy.bugs.groovy9204
+            @groovy.transform.CompileStatic
+            class ArrayListTest extends Three {
+                def test() {
+                    field = new ArrayList()
+                    field.add("hello")
+                    field[0]
+                }
+            }
+
+            @groovy.transform.CompileStatic
+            class LinkedListTest extends Four {
+                def test() {
+                    field = new LinkedList()
+                    field.addFirst("hello")
+                    field[0]
+                }
+            }
+            
+            assert new ArrayListTest().test() == 'hello'
+            assert new LinkedListTest().test() == 'hello'
+        '''
+    }
+}
